meta: Port over to new dinit android-service using arguments
Now we no longer explicitly enable any halium-wrappers dinit service but
these are purely created on runtime similar to the original systemd
implementation of them by Droidian.

For example we can just "dinitctl start android-service@hwcomposer" with
the pre-existing configuration in /etc/android-service/hwcomposer.env
which is additionally overrideable on a per-port basis if the defaults
don't work for some reason.